Changements mineurs
Gestion d'erreurs éventuelles

+++ Partie bash/import-archive.sh
@@ -8,7 +8,7 @@ forcer=0;
 for fichier in "$@"
 do
     #bonus: option -f avec deux paramètres = erreur
-    if [ $fichier == "-f" ]; then
+    if [ "$fichier" == "-f" ]; then
         if [ $# -ne 2 ]; then echo "Un seul paramètre avec l'option -f"; exit 5; fi
         
         forcer=1;
@@ -31,13 +31,20 @@ do
 
         #ajout d'un fichier -> incrémentation de N
         N=$(($N+1));
+
         #copie du fichier initial sauf le chiffre N au début dans un fichier tmp
         tail -n +2 "$REPERTOIRE/archives" > "$nom_fichier.tmp"
+        if [ $? -ne 0 ]; then echo "Un problème a eu lieu lors de la copie"; exit 4; fi
+
         #fait rentrer le nouveau N au tout début du fichier (fichier vide initialement donc > suffit)
         echo "$N" > "$REPERTOIRE/archives"
+
         # >> c'est le contenu du premier fichier + le contenu du 2eme fichier (fichier non vide donc concaténation obligatoire >>)
         cat "$nom_fichier.tmp" >> "$REPERTOIRE/archives"
+        if [ $? -ne 0 ]; then echo "Un problème a eu lieu lors de la copie"; exit 4; fi
+
         rm "$nom_fichier.tmp"
+        if [ $? -ne 0 ]; then echo "Un problème a eu lieu lors de la copie"; exit 4; fi
 
         #date d'ajout du fichier dans l'archive = date correspondante à l'exécution du code
         date_ajout=$(date +"%Y%m%d-%H%M%S");
@@ -45,17 +52,21 @@ do
         echo "$nom_fichier:$date_ajout:" >> "$REPERTOIRE/archives"
     else
         #passage en force donc on copie quitte à écraser un fichier avec un nom identique
-        if [ $forcer -eq 1 ]; then
+        if [ "$forcer" -eq 1 ]; then
+            #littéralement un copier / coller du code 10 lignes au dessus
             cp "$chemin" "$REPERTOIRE/"
             if [ $? -ne 0 ]; then echo "Un soucis a eu lieu lors de la copie du fichier dans $REPERTOIRE."; exit 3; fi
 
-            #copie du fichier initial sauf le chiffre N au début dans un fichier tmp
             tail -n +2 "$REPERTOIRE/archives" > "$nom_fichier.tmp"
-            #fait rentrer le nouveau N au tout début du fichier (fichier vide initialement donc > suffit)
+            if [ $? -ne 0 ]; then echo "Un problème a eu lieu lors de la copie"; exit 4; fi
+
             echo "$N" > "$REPERTOIRE/archives"
-            # >> c'est le contenu du premier fichier + le contenu du 2eme fichier (fichier non vide donc concaténation obligatoire >>)
+
             cat "$nom_fichier.tmp" >> "$REPERTOIRE/archives"
+            if [ $? -ne 0 ]; then echo "Un problème a eu lieu lors de la copie"; exit 4; fi
+
             rm "$nom_fichier.tmp"
+            if [ $? -ne 0 ]; then echo "Un problème a eu lieu lors de la copie"; exit 4; fi
 
             #date d'ajout du fichier dans l'archive = date correspondante à l'exécution du code
             date_ajout=$(date +"%Y%m%d-%H%M%S");
@@ -65,7 +76,7 @@ do
 
         else
             echo "Le fichier $nom_fichier existe déjà dans l'archive, voulez-vous l'écraser ? (oui ou non)";
-            read verification;
+            read -r verification;
 
             if [ "$verification" == "oui" ]; then
                 cp "$chemin" $REPERTOIRE/;
